split up cell type distribution chart components for readability, add titles/tooltips to expanded graphs
